pyproject.toml
[tool.poetry]
name = "nba-scorigami"
version = "0.1.0"
description = ""
authors = ["svruddar <svrdev26@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.1.4"
bs4 = "^0.0.1"
html5lib = "^1.1"
mysql-connector-python = "^8.2.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
